+++ CHANGELOG.md
@@ -19,6 +19,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
+## [1.1.7] - 2026-03-09
+
+### Changed
+
+- Version bump.
+
+---
+
 ## [1.1.6] - 2026-03-08
 
 ### Changed